- Fix Open in EMM not working if the path has space - The default folder of EMM is now Macroes when open - Add confirm dialog when trying to delete macro in AEMG-EX

+++ AEMG-EX.Core/Concretes/AEScriptGenerator.cs
@@ -115,8 +115,8 @@ namespace AEMG_EX.Core
 
             try
             {
-                GlobalData.ScaleX = setting.CustomX / macro.OriginalX;
-                GlobalData.ScaleY = setting.CustomY / macro.OriginalY;
+                GlobalData.ScaleX = (double)setting.CustomX / macro.OriginalX;
+                GlobalData.ScaleY = (double)setting.CustomY / macro.OriginalY;
             }
             catch
             {
